requirements.txt
ncclient==0.5.3
netmiko==1.4.2
PyYAML==3.12
requests==2.18.2
xmltodict==0.11.0
flask
